+++ src/main/java/br/com/renanlabs/mvc/financesonpoint/service/OperacaoBasica.java
@@ -1,9 +1,11 @@
 package br.com.renanlabs.mvc.financesonpoint.service;
 
+import java.math.BigDecimal;
+
 import br.com.renanlabs.mvc.financesonpoint.model.Carteira;
 
 public interface OperacaoBasica {
 
-	public void efetua(Carteira carteira, Double valor);
+	public void efetua(Carteira carteira, BigDecimal valor);
 	
 }
